requirements.txt
aio-pika==9.5.5
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiomysql==0.1.1
aioredis==1.3.1
aiormq==6.8.1
aiosignal==1.3.2
alembic==1.15.2
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
asyncpg==0.30.0
attrs==25.3.0
Authlib==1.6.0
bcrypt==4.3.0
beautifulsoup4==4.13.4
billiard==4.2.1
boto3==1.38.36
botocore==1.38.36
celery==5.5.3
certifi==2025.4.26
cffi==1.17.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==45.0.4
dnspython==2.7.0
dynaconf==3.2.11
email_validator==2.2.0
et_xmlfile==2.0.0
exceptiongroup==1.3.0
fastapi==0.115.12
fastapi-cache2==0.2.2
fastapi-cli==0.0.7
fastapi-utils==0.8.0
frozenlist==1.6.0
greenlet==3.2.1
gunicorn==23.0.0
h11==0.14.0
hiredis==3.2.1
httpcore==0.17.3
httptools==0.6.4
httpx==0.24.1
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
kombu==5.5.4
lxml==5.4.0
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.4.3
mypy_extensions==1.1.0
openpyxl==3.1.5
orjson==3.10.18
packaging==25.0
pamqp==3.3.0
passlib==1.7.4
pendulum==3.1.0
pluggy==1.6.0
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.22.1
prompt_toolkit==3.0.51
propcache==0.3.1
psutil==5.9.8
pycparser==2.22
pydantic==2.11.7
pydantic-extra-types==2.10.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.19.1
PyJWT==2.10.1
PyMySQL==1.1.1
pytest==8.4.0
pytest-asyncio==1.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
redis==4.6.0
rich==14.0.0
rich-toolkit==0.14.4
s3transfer==0.13.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
starlette==0.46.2
transitions==0.9.2
typer==0.15.3
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
ujson==5.10.0
urllib3==2.4.0
uvicorn==0.21.1
vine==5.1.0
watchfiles==1.0.5
wcwidth==0.2.13
websockets==15.0.1
yarl==1.20.0
